requirements.txt
paho-mqtt==1.1
pep8==1.6.2
protobuf==2.6.1
